Improved commander buff with particles #298
