+++ src/main/java/net/shadowmage/ancientwarfare/npc/entity/faction/NpcFactionMountedSoldier.java
@@ -5,7 +5,7 @@ import net.minecraft.entity.ai.*;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.world.World;
 import net.shadowmage.ancientwarfare.npc.ai.*;
-import net.shadowmage.ancientwarfare.npc.ai.faction.NpcAIFactionFindCommander;
+import net.shadowmage.ancientwarfare.npc.ai.faction.NpcAIFactionCommander;
 import net.shadowmage.ancientwarfare.npc.ai.faction.NpcAIFactionRideHorse;
 
 public abstract class NpcFactionMountedSoldier extends NpcFactionMounted {
@@ -18,7 +18,7 @@ public abstract class NpcFactionMountedSoldier extends NpcFactionMounted {
         this.tasks.addTask(0, new EntityAIRestrictOpenDoor(this));
         this.tasks.addTask(0, new EntityAIOpenDoor(this, true));
         this.tasks.addTask(0, (horseAI = new NpcAIFactionRideHorse(this)));
-        this.tasks.addTask(1, new NpcAIFactionFindCommander(this));
+        this.tasks.addTask(1, new NpcAIFactionCommander(this));
         this.tasks.addTask(1, new NpcAIFollowPlayer(this));
         this.tasks.addTask(2, new NpcAIMoveHome(this, 50F, 5F, 30F, 5F));
         this.tasks.addTask(3, new NpcAIAttackMeleeLongRange(this));
